Ensure that groups don't have multiple parents

+++ lib/synx/pbx_variant_group.rb
@@ -5,10 +5,13 @@ module Xcodeproj
     module Object
       class PBXVariantGroup
         
-        def sync
+        def sync(group)
+          ensure_internal_consistency(group)
           folder_path = children.first.real_path.parent
           work_destination_pathname = parent.work_pathname
-          FileUtils.mv(folder_path, work_destination_pathname.realpath)
+          if folder_path.exist?
+            FileUtils.mv(folder_path, work_destination_pathname.realpath)
+          end
         end
 
       end
